Refactor API: Add PATCH endpoint to decrease product quantity (#1234)
This commit refactors the API by adding a PATCH endpoint to decrease the
product quantity. The endpoint is implemented in the `cms_product`
controller and can be accessed using the `/cms_product/{id}` route. This
change enables the application to handle product quantity updates more
efficiently and accurately. It also introduces a new
`CmsProductProducer` dependency in the `CmsProductController` to support
message queuing for further processing. The existing code has been
modified to accommodate these changes.

Issue: #1234